requirements.txt
# Python dependencies for IDXR matching algorithms
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1

# Data Processing
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0

# ML/AI
tensorflow==2.14.0
torch==2.1.1
transformers==4.35.2

# Utilities
httpx==0.25.1
aiofiles==23.2.1
python-dateutil==2.8.2
pytz==2023.3
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1

# Monitoring
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0